# docs/archive/openapi.monolith.yaml
openapi: 3.0.3
info:
  title: Vouchstone
  version: 0.1.0
  description: >-
    Evidenced cyber assurance. Vouchstone caps every domain posture score by the dated
    evidence behind it, decays scores when that evidence lapses, records who actually
    discovered each successful attack, reconciles believed impact against actual control
    spend, and generates the committee pack, regulatory return, and insurance proposal
    response from one traceable evidence base.
  x-product-slug: ai-cyber-vouchstone
  x-source: ai-in-cyber/Accenture-Security-Report-2016_Key-Insights_SPAIN-esp/
servers:
  - url: https://api.example.com
...
paths:
...
  /v1/evidence:
    get:
      operationId: listEvidenceItems
      tags: [Evidence]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: domain
          in: query
          schema:
            $ref: "#/components/schemas/DomainKey"
        - name: validity
          in: query
          schema:
            type: string
            enum: [valid, expiring, expired, superseded]
      responses:
        "200":
          description: Evidence items page with validity state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeEvidenceItem"
    post:
      operationId: submitEvidenceItem
      tags: [Evidence]
      description: >-
        Submit typed evidence. The validity period is derived from the committee-owned
        rule for the evidence type and cannot be set by the submitter.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceItemCreate"
      responses:
        "201":
          description: Evidence accepted with a derived expiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeEvidenceItem"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /v1/exercises:
    get:
      operationId: listEmulationExercises
      tags: [Exercises]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, executed, deferred, missed, expired]
      responses:
        "200":
          description: Exercises page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeEmulationExercise"
    post:
      operationId: scheduleEmulationExercise
      tags: [Exercises]
      description: >-
        Schedule an independent adversary emulation with a defined scope over named
        priority assets. A missed due date expires the associated evidence.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmulationExerciseCreate"
      responses:
        "201":
          description: Exercise scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeEmulationExercise"
...
  /v1/findings:
    get:
      operationId: listFindings
      tags: [Findings]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: state
          in: query
          schema:
            type: string
            enum: [open, overdue, closed, accepted_risk]
        - name: domain
          in: query
          schema:
            $ref: "#/components/schemas/DomainKey"
      responses:
        "200":
          description: Findings page with suppressed domain scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeFinding"
    post:
      operationId: raiseFinding
      tags: [Findings]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindingCreate"
      responses:
        "201":
          description: Finding raised with a named owner and due date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeFinding"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
...
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    DomainKey:
      type: string
      enum:
        - risk_exposure
        - governance_and_leadership
        - strategic_context
        - resilience
        - response_capability
        - extended_ecosystem
        - efficient_investment
...
    EvidenceItem:
      type: object
      required: [id, typeKey, domains, effectiveDate, validity]
      properties:
        id:
          type: string
        typeKey:
          type: string
        domains:
          type: array
          items:
            $ref: "#/components/schemas/DomainKey"
        assetIds:
          type: array
          items:
            type: string
        effectiveDate:
          type: string
          format: date
        expiresAt:
          type: string
          format: date-time
        validity:
          type: string
          enum: [valid, expiring, expired, superseded]
        provenance:
          type: string
          enum: [internal, independent_provider, regulator, supplier, auditor]
        providerName:
          type: string
        artefactReference:
          type: string
        supersededById:
          type: string
    EvidenceItemCreate:
      type: object
      required: [typeKey, domains, effectiveDate, provenance]
      properties:
        typeKey:
          type: string
        domains:
          type: array
          items:
            $ref: "#/components/schemas/DomainKey"
        assetIds:
          type: array
          items:
            type: string
        effectiveDate:
          type: string
          format: date
        provenance:
          type: string
          enum: [internal, independent_provider, regulator, supplier, auditor]
        providerName:
          type: string
        artefactReference:
          type: string
...
    EmulationExercise:
      type: object
      required: [id, status, dueDate]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [scheduled, executed, deferred, missed, expired]
        dueDate:
          type: string
          format: date
        executedAt:
          type: string
          format: date-time
        providerName:
          type: string
        providerIndependent:
          type: boolean
        scopeAssetIds:
          type: array
          items:
            type: string
        scenario:
          type: string
          enum: [insider_privilege_abuse, determined_external_adversary, supply_chain_entry, ransomware_resilience, data_exfiltration]
        outcome:
          type: string
          enum: [objectives_blocked, partially_blocked, objectives_achieved, inconclusive]
        detectedByDefenders:
          type: boolean
        findingIds:
          type: array
          items:
            type: string
        evidenceItemId:
          type: string
    EmulationExerciseCreate:
      type: object
      required: [dueDate, providerName, scopeAssetIds, scenario]
      properties:
        dueDate:
          type: string
          format: date
        providerName:
          type: string
        providerIndependent:
          type: boolean
          default: true
        scopeAssetIds:
          type: array
          items:
            type: string
        scenario:
          type: string
          enum: [insider_privilege_abuse, determined_external_adversary, supply_chain_entry, ransomware_resilience, data_exfiltration]
...
    FindingCreate:
      type: object
      required: [title, severity, domains, ownerId, dueDate]
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        domains:
          type: array
          items:
            $ref: "#/components/schemas/DomainKey"
        assetIds:
          type: array
          items:
            type: string
        ownerId:
          type: string
        dueDate:
          type: string
          format: date
        origin:
          type: string
          enum: [exercise, incident, evidence_lapse, audit, allocation_divergence, coverage_gap]
    Finding:
      type: object
      required: [id, title, severity, state, ownerId, dueDate]
      properties:
        id:
          type: string
        title:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        state:
          type: string
          enum: [open, overdue, closed, accepted_risk]
        domains:
          type: array
          items:
            $ref: "#/components/schemas/DomainKey"
        assetIds:
          type: array
          items:
            type: string
        ownerId:
          type: string
        dueDate:
          type: string
          format: date
        origin:
          type: string
        suppressingDomainScore:
          type: boolean
        remediation:
          $ref: "#/components/schemas/RemediationCommitment"
    RemediationCommitment:
      type: object
      required: [ownerId, targetDate]
      properties:
        ownerId:
          type: string
        targetDate:
          type: string
          format: date
        plan:
          type: string
        revisionCount:
          type: integer
        acceptedRiskRationale:
          type: string
        closedAt:
          type: string
          format: date-time
        closureApprovedBy:
          type: string
...
    DataEnvelopeEvidenceItem:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/EvidenceItem"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeEmulationExercise:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/EmulationExercise"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    DataEnvelopeFinding:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Finding"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeEvidenceItem:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/EvidenceItem"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeEmulationExercise:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/EmulationExercise"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ListEnvelopeFinding:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Finding"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...