# platform/api-server/src/domains/evidence/handlers/evidence.handlers.ts
/**
 * Handlers for Evidence
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: evidence
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { EvidenceDomainModule } from "../dependencies/evidence-ddd.dependencies.js";
import type { ListEvidenceItemsInput, SubmitEvidenceItemInput } from "@vouchstone/services/evidence";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/evidence
 * Operation: listEvidenceItems
 */
export async function listEvidenceItems(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], domain?: string | string[], validity?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: EvidenceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListEvidenceItemsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.evidences.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/evidence
 * Operation: submitEvidenceItem
 */
export async function submitEvidenceItem(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: EvidenceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: SubmitEvidenceItemInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.evidences.create.execute(input);
  return reply.code(201).send(result);
}