# platform/api-server/src/domains/exercises/handlers/exercise.handlers.ts
/**
 * Handlers for Exercise
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: exercises
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ExercisesDomainModule } from "../dependencies/exercises-ddd.dependencies.js";
import type { ListEmulationExercisesInput, ScheduleEmulationExerciseInput } from "@vouchstone/services/exercises";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/exercises
 * Operation: listEmulationExercises
 */
export async function listEmulationExercises(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], status?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: ExercisesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListEmulationExercisesInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.exercises.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/exercises
 * Operation: scheduleEmulationExercise
 */
export async function scheduleEmulationExercise(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: ExercisesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ScheduleEmulationExerciseInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.exercises.get.execute(input);
  return reply.code(201).send(result);
}