# platform/api-server/src/domains/findings/handlers/finding.handlers.ts
/**
 * Handlers for Finding
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: findings
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { FindingsDomainModule } from "../dependencies/findings-ddd.dependencies.js";
import type { ListFindingsInput, RaiseFindingInput } from "@vouchstone/services/findings";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/findings
 * Operation: listFindings
 */
export async function listFindings(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], state?: string | string[], domain?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: FindingsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListFindingsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.findings.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/findings
 * Operation: raiseFinding
 */
export async function raiseFinding(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: FindingsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: RaiseFindingInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.findings.get.execute(input);
  return reply.code(201).send(result);
}